Core –- Parts and Usage Overview

The code is divided up into files to contain pieces with similar purposes or concepts in the algorithm. Each file has its own single module for defining a namespace used when importing its names into other files. Each module exports members intended for public access, but the code in this project explicitly names its imports to maintain clarity in what is used and where it comes from.

Maps.jl

InformativeSampling.Maps — Module

This module contains types for holding and handling real and synthetic data. Mainly used by, but foundational for the other parts too.

Main public types and functions:

Bounds

Map

cellToPoint

generateAxes

getBounds

pointToCell

randomPoint

res

InformativeSampling.Maps.Bounds — Type

struct NamedTuple{(:lower, :upper), Tuple{Vector{Float64}, Vector{Float64}}}

The bounds of the region. Consists of the lower and upper bounds, each a list of floating-point values.

InformativeSampling.Maps.Map — Type

A general type for holding multidimensional data (usually a matrix) along with associated dimension bounds. It's main purpose is to handle the conversion between world coordinates and grid indices internally. Converting between the two representations treats rows as the first variable (x-axis), columns as the second (y-axis), and so on.

Its typical use is to act as a 2D map of some value that can be sampled. A Map will return the value of the grid cell that a given point falls within. In other words, the map value is constant within each cell. One can also think of this as a nearest-neighbor approximation.

Each cell index is treated as the center of its cell. Thus the map's lower bound is at the center of the first cell and the map's upper bound is at the center of the last cell.

Also made to function directly like a built-in N-dimensional array by sub-typing and implementing the base methods.

Fields:

data::AbstractArray: N-dimensional array of data

bounds::@NamedTuple{lower::Vector{Float64}, upper::Vector{Float64}}: vectors of lower and upper bounds, defaults to zeros and ones

InformativeSampling.Maps.Map — Method

Method accepts a single vector (the location), returns a scalar (the value at that point).

Examples

data = reshape(1:25, 5, 5)
bounds = (
    lower = [0.0, 0.0],
    upper = [1.0, 1.0]
)
m = Map(data, bounds)
m2 = Map(data) # bounds will be zero to one

x = [.2, .75]
val = m(x) # returns the value at a single 2D point
val2 = m[1,4] # can also use as if it's just the underlying matrix

Base.rand — Method

rand(map::InformativeSampling.Maps.Map) -> Tuple{Any, Any}

Generates a random point in the map. Returns the location and its value.

Examples

data = reshape(1:25, 5, 5)
map = Map(data)
rand(map)

InformativeSampling.Maps.cellToPoint — Method

cellToPoint(ci, map) -> Any

Takes in a CartesianIndex and a Map and returns a point in world-coordinates.

InformativeSampling.Maps.checkBounds — Method

checkBounds(x, map::InformativeSampling.Maps.Map) -> Any

Function emits error if location is outside of map bounds.

Examples

x = [.2, .75]
data = reshape(1:25, 5, 5)
map = Map(data)
checkBounds(x, map) # no error thrown

InformativeSampling.Maps.generateAxes — Method

generateAxes(map) -> Tuple{Any, Any}

Method to generate the x, y, etc. axes and points of a Map. Useful for plotting.

InformativeSampling.Maps.getBounds — Method

Get the lower and upper bounds of the map.

InformativeSampling.Maps.pointToCell — Method

pointToCell(x, map) -> Any

Takes in a point in world-coordinates and a Map and returns a CartesianIndex for the underlying array.

InformativeSampling.Maps.randomPoint — Method

randomPoint(map::InformativeSampling.Maps.Map) -> Any

Generates a random point in the map. Returns the location.

InformativeSampling.Maps.res — Method

res(map) -> Any

Returns the resolution for each dimension of the given Map as a vector.

Missions.jl

InformativeSampling.Missions — Module

This module contains functions for initializing mission data and the function for running the entire search mission. The entry-point to the actual informative sampling. This contains the main loop and most of the usage of Samples and belief models.

Main public types and functions:

Mission

replay

InformativeSampling.Missions.Mission — Type

The main function that runs the informative sampling routine. For each iteration, a sample location is selected, a sample is collected, the belief model is updated, and visuals are possibly shown. The run finishes when the designated number of samples is collected.

Inputs:

func: any function to be run at the end of the update loop, useful for visualization or saving data (default does nothing)

samples: a vector of samples, this can be used to jump-start a mission or resume a previous mission (default empty)

beliefs: a vector of beliefs, this pairs with the previous argument (default empty)

seed_val: the seed for the random number generator, an integer (default 0)

sleep_time: the amount of time to wait after each iteration, useful for visualizations (default 0)

Outputs:

samples: a vector of new samples collected

beliefs: a vector of probabilistic representations of the quantities being searched for, one for each sample collection

Examples

using Missions: synMission

mission = synMission(num_samples=10) # create the specific mission
samples, beliefs = mission(visuals=true, sleep_time=0.5) # run the mission

InformativeSampling.Missions.Mission — Type

Fields:

occupancy::Any: an occupancy map, true in cells that are occupied

sampler::Any: a function that returns a measurement value for any input

num_samples::Any: the number of samples to collect in one run

sampleCostType::Any: a constructor for the function that returns the (negated) value of taking a sample (default DistScaledEIGF)

weights::Any: weights for picking the next sample location

start_locs::Any: the locations that should be sampled first (default [])

prior_samples::Any: any samples taken previously (default empty)

kernel::Any: the kernel to be used in the belief model (default multiKernel)

means::Any: whether or not to use a non-zero mean for each quantity and to learn means (default (true, false))

noise::Any: a named tuple of noise value(s) and if learned further (default (0.0, false))

use_cond_pdf::Any: whether or not to use the conditional distribution of the data to train the belief model (default false)

hyp_drop::Any: whether or not to drop hypotheses and settings for it (default (false, 10, 5, 0.4))

Defined as a keyword struct, so all arguments are passed in as keywords:

mission = Mission(; occupancy,
                  sampler,
                  num_samples,
                  sampleCostType,
                  weights,
                  start_locs,
                  prior_samples,
                  noise,
                  kernel)

InformativeSampling.Missions.replay — Method

replay(
    func,
    M::InformativeSampling.Missions.Mission,
    full_samples,
    beliefs;
    sleep_time
)

Replays a mission that has already taken place. Mainly for visualization purposes.

Inputs:

func: any function to be run at the end of the update loop, useful for

visualization or saving data (default does nothing)

full_samples: a vector of samples

beliefs: a vector of beliefs

sleep_time: the amount of time to wait after each iteration, useful for

visualizations (default 0)

MultiQuantityGPs.jl

InformativeSampling.MultiQuantityGPs — Module

This module contains everything to do with what is inferred about values in the environment. In practical terms: means, variances, and correlations. This is all built on Gaussian Processes.

Main public types and functions:

MQGP

fullCov

meanDerivAndVar

quantityCorMat

quantityCovMat

InformativeSampling.MultiQuantityGPs.LinearModel — Type

A multivariate linear model of $y$ dependent on $x$ with parameters $a$ and $b$ of the form

\[Y = a + b^T X\]

X and Y are matrices containing the points as columns.

InformativeSampling.MultiQuantityGPs.LinearModel — Method

Returns a linear model of set of variables Y conditioned on set X. Requires full mean vector and covariance matrix of the joint normal distribution.

InformativeSampling.MultiQuantityGPs.MQGP — Type

struct MQGP{T}

Belief model struct and function for multiple quantities with 2D inputs.

Designed on top of a Multi-Quantity Gaussian Process, but can still be used with a single quantity.

Its interface: X -> μ, σ (SampleInputs -> means, standard deviations)

InformativeSampling.MultiQuantityGPs.MQGP — Method

MQGP(
    samples,
    bounds::@NamedTuple{lower::Vector{Float64}, upper::Vector{Float64}};
    N,
    kernel,
    means,
    noise,
    use_cond_pdf
) -> InformativeSampling.MultiQuantityGPs.MQGP{typeof(InformativeSampling.Kernels.multiKernel)}

Creates and returns a MQGP with hyperparameters trained and conditioned on the samples given. Lower and upper bounds are used to initialize one of the hyperparameters.

A noise standard deviation can optionally be passed in either as a single scalar value for all samples or a vector of values, one for each sample.

Examples

# create a MQGP
beliefModel = MQGP([M.prior_samples; samples], bounds)

InformativeSampling.MultiQuantityGPs.MQGP — Method

Inputs:

X: a single sample input or an array of multiple

full_cov: (optional) if this is true, returns the full covariance matrix in place of the vector of standard deviations

Outputs:

μ, σ: a pair of expected value(s) and uncertainty(s) for the given point(s)

Examples

X = [([.1, .2], 1),
     ([.2, .1], 2)]
μ, σ = beliefModel(X) # result: [μ1, μ2], [σ1, σ2]

InformativeSampling.MultiQuantityGPs.createLossFunc — Method

createLossFunc(
    X,
    Y_vals,
    Y_errs,
    kernel,
    use_cond_pdf
) -> InformativeSampling.MultiQuantityGPs.var"#15#16"

This function creates the loss function for training the GP. The negative log marginal likelihood is used.

InformativeSampling.MultiQuantityGPs.fullCov — Method

fullCov(
    bm::InformativeSampling.MultiQuantityGPs.MQGP,
    X::AbstractArray{Tuple{Vector{Float64}, Int64}}
) -> Any

Returns the full covariance matrix for the belief model.

InformativeSampling.MultiQuantityGPs.latentCovMat — Method

latentCovMat(
    bm::InformativeSampling.MultiQuantityGPs.MQGP{typeof(InformativeSampling.Kernels.multiKernel)}
) -> Any

Gives the covariance matrix between all latent functions from the hyperparameters.

InformativeSampling.MultiQuantityGPs.meanDerivAndVar — Method

meanDerivAndVar(
    bm::InformativeSampling.MultiQuantityGPs.MQGP,
    x::Tuple{Vector{Float64}, Int64}
) -> Any

Returns the normed gradient of the mean of the belief model and its variance.

InformativeSampling.MultiQuantityGPs.optimizeLoss — Method

optimizeLoss(lossFunc, θ0; solver, iterations) -> Any

Routine to optimize the lossFunc and return the optimal parameters θ.

Can pass in a different solver. NelderMead is picked as default for better speed with about the same performance as LFBGS.

InformativeSampling.MultiQuantityGPs.quantityCorMat — Method

quantityCorMat(beliefModel::MQGP)

Gives the correlation matrix between all quantities from the hyperparameters.

InformativeSampling.MultiQuantityGPs.quantityCovMat — Method

quantityCovMat(
    bm::InformativeSampling.MultiQuantityGPs.MQGP
) -> Any

Gives the covariance matrix between all quantities from the hyperparameters. The model of the quantities is the latent functions plus their measurement noise.

Samples.jl

InformativeSampling.Samples — Module

This module contains everything to do with sampling values in the environment. Its alias types Location and SampleInput are fundamental pieces for MultiQuantityGPs as well.

Main public types and functions:

Location

MapsSampler

Sample

SampleInput

UserSampler

selectSampleLocation

takeSamples

InformativeSampling.Samples.Location — Type

mutable struct Array{Float64, 1} <: DenseVector{Float64}

Location of sample

InformativeSampling.Samples.MapsSampler — Type

Handles samples of the form (location, quantity) to give the value from the right map. Internally a tuple of Maps.

Constructor can take in a tuple or vector of Maps or each Map as a separate argument.

Examples

ss = MapsSampler(Map(zeros(5, 5)), Map(ones(5, 5)))

loc = [.2, .75]
ss(loc) # result: [0, 1]
ss((loc, 2)) # result: 1

InformativeSampling.Samples.Sample — Type

Struct to hold the input and output of a sample.

Fields:

x::Tuple{Vector{Float64}, Int64}: the sample input, usually a location and sensor id

y::Any: the sample output or observation, a scalar

InformativeSampling.Samples.SampleInput — Type

struct Tuple{Vector{Float64}, Int64}

Sample input, the combination of: (Location, sensor index)

InformativeSampling.Samples.SampleOutput — Type

struct Tuple{Float64, Float64}

Value of sample measurement, the measurement mean and standard deviation

InformativeSampling.Samples.UserSampler — Type

A sampler that asks the user to input measurement values, one for each quantity at the given location.

InformativeSampling.Samples.selectSampleLocation — Method

selectSampleLocation(sampleCost, bounds) -> AbstractArray

The optimization of choosing a best single sample location.

Inputs:

sampleCost: a function from sample location to cost (x->cost(x))

bounds: map lower and upper bounds

Returns the sample location, a vector

InformativeSampling.Samples.takeSamples — Method

takeSamples(loc, sampler) -> Any

Pulls a ground truth value from a given location and constructs a Sample object to hold them both.

Inputs:

loc: the location to sample

sampler: a function that returns ground truth values

quantities: (optional) a vector of integers which represent which quantities to sample, defaults to all of them

Outputs a vector of Samples containing input x and measurement y

SampleCosts.jl

InformativeSampling.SampleCosts — Module

This module holds a variety of SampleCost functions used by Samples.jl in selecting a new sample location. The purpose is to pick the location that will minimize the given function.

Each sample cost function in this module is a subtype of the abstract SampleCost type. Their common interface consists of two functions:

values(sampleCost, location): returns the values of the terms (μ, σ, τ, P);
this is typically what each subtype will override;
in all of these, μ = belief model mean, σ = belief model std, τ = travel distance, P = proximity value;
all of these are for the specific location

sampleCost(location): the actual sample cost at the location; it has a default method explained in SampleCost;
for the equations the location is denoted $x$

Many of these were experiment cost functions and aren't recommended. The main ones recommended for use, in order, are

DistScaledEIGF

EIGF

MIPT

OnlyVar

Others can be useful if one wants to do some more experimentation. Note that all these functions are currently hardcoded to use the first quantity as the objective quantity unless otherwise stated. Unless explicitly used, the distance value is generally used to make sure unreachable locations are forbidden (their value will be Inf).

Main public types and functions:

DerivVar

DistLogEIGF

DistProx

DistScaledDerivVar

DistScaledEIGF

EIGF

InfoGain

LogLikelihood

LogLikelihoodFull

LogNormed

MIPT

OnlyVar

SampleCost

VarTrace

values

InformativeSampling.SampleCosts.DerivVar — Type

Uses the norm of the derivative of the belief model mean and the belief model variance:

\[C(x) = - w_1 \, {\left\lVert \frac{\partial μ}{\partial x}(x) \right\rVert}^2 - w_2 \, σ^2(x)\]

InformativeSampling.SampleCosts.DistLogEIGF — Type

A variation on EIGF that takes the logarithm of the variance and adds a distance cost term that is normalized by the average of the region dimensions:

\[C(x) = - w_1 \, (μ(x) - y(x_c))^2 - w_2 \, \log(σ^2(x)) + w_3 \, β \, \frac{τ(x)}{\left\lVert \boldsymbol{\ell}_d \right\rVert_1}\]

where $β$ is a parameter to delay the distance effect until a few samples have been taken.

InformativeSampling.SampleCosts.DistProx — Type

Combines the average mean value, average standard deviation, travel distance, and proximity as terms:

\[C(x) = - w_1 \, μ_{\mathrm{ave}}(x) - w_2 \, σ_{\mathrm{ave}}(x) + w_3 \, τ(x) + w_4 \, P(x)\]

where $P(x) = \sum_i(\frac{\min(\boldsymbol{\ell}_d)}{4 \, \mathrm{dist}_i})^3$. Averages are performed over all quantities.

InformativeSampling.SampleCosts.DistScaledDerivVar — Type

Uses the norm of the derivative of the belief model mean and the belief model variance, then scales it all by a normalized travel distance:

\[C(x) = \frac{- w_1 \, {\left\lVert \frac{\partial μ}{\partial x}(x) \right\rVert}^2 - w_2 \, σ^2(x)} {1 + β \, \frac{τ(x)}{\left\lVert \boldsymbol{\ell}_d \right\rVert_1}}\]

where $β$ is a parameter to delay the distance effect until a few samples have been taken.

InformativeSampling.SampleCosts.DistScaledEIGF — Type

Augments EIGF with a factor to scale by a normalized travel distance:

\[C(x) = \frac{- w_1 \, (μ(x) - y(x_c))^2 - w_2 \, σ^2(x)} {1 + β \, \frac{τ(x)}{\left\lVert \boldsymbol{\ell}_d \right\rVert_1}}\]

where $β$ is a parameter to delay the distance effect until a few samples have been taken.

InformativeSampling.SampleCosts.EIGF — Type

Expected Informativeness for Global Fit (EIGF). This function is adapted from [Lam] through added weights to choose the balance between exploration and exploitation. It has the form:

\[C(x) = - w_1 \, (μ(x) - y(x_c))^2 - w_2 \, σ^2(x)\]

where $x_c$ is the nearest collected sample location.

InformativeSampling.SampleCosts.InfoGain — Type

Derived from the idea of information gain across the region. Returns the entropy of a set of points (10x10 grid) given the new sample location. Minimizing this entropy is equivalent to maximizing information gain since the entropy before the sample is always the same.

This function is very computationally expensive, which is why the test grid is set at 10x10.

It has the form:

\[C(x) = \log |Σ|\]

InformativeSampling.SampleCosts.LogLikelihood — Type

Idea derived from the log likelihood of query location. Similar to EIGF, the measured value from the nearest sample is used:

\[C(x) = - w_1 \, \left( \frac{μ(x) - y(x_c)}{σ_n} \right)^2 - w_2 \, \log (σ^2(x))\]

where $x_c$ is the nearest collected sample location, and $σ_n$ is the noise.

This function was seen to work well but hasn't undergone extensive tests. There is still some question about the theory and the use of noise parameter vs signal amplitude parameter.

InformativeSampling.SampleCosts.LogLikelihoodFull — Type

A test of the log likelihood idea but using a weighted sum of all measured sample values, not just the nearest one:

\[C(x) = - w_1 \, \frac{1}{\sum_i k(x, x_i)} \sum_i k(x, x_i) * \left( \frac{μ(x) - y(x_i)}{σ_n} \right)^2 - w_2 \, \log (σ^2(x))\]

where $x_i$ is each collected sample location, and $σ_n$ is the noise.

This function's performance wasn't satisfactory.

InformativeSampling.SampleCosts.LogNormed — Type

Combines the average belief value and the log of the average uncertainty value of all quantities. All belief and uncertainty values are first normalized by the maximum belief value of that quantity. It has the form:

\[C(x) = - w_1 \, μ_{\textrm{norm-ave}}(x) - w_2 \, \log (σ_{\textrm{norm-ave}}(x))\]

InformativeSampling.SampleCosts.MEPE — Type

This was a cost function being considered but wasn't finished. Taken from [Liu].

InformativeSampling.SampleCosts.MIPT — Type

A simple cost function that doesn't use a belief model but works purely on distances. It returns the negated distance to the nearest sample:

\[C(x) = - \min_i \left\lVert x - x_i \right\rVert\]

Useful when maximizing distance between samples.

InformativeSampling.SampleCosts.OnlyVar — Type

Returns the negated variance of the belief model at the query location:

\[C(x) = - σ^2\]

InformativeSampling.SampleCosts.SampleCost — Type

Typically construct a SampleCost through SampleCostType(occupancy, samples, beliefModel, quantities, weights)

A pathCost is constructed automatically from the other arguments.

This object can then be called to get the cost of sampling at a location: sampleCost(x)

InformativeSampling.SampleCosts.SampleCost — Method

Cost to take a new sample at a location $x$. This is a fallback method that calculates a simple linear combination of all the values of a SampleCost.

Has the form:

\[C(x) = - w_1 \, μ(x) - w_2 \, σ(x) + w_3 \, τ(x) + w_4 \, P(x)\]

InformativeSampling.SampleCosts.VarTrace — Type

Similar to InfoGain but uses only variances rather than the full covariance matrix, which is the same as the trace instead of the log determinant. This reduces computation but is still more costly than a single point estimate. This uses a 20x20 grid of test points.

It has the form:

\[C(x) = \textrm{tr}(Σ)\]

InformativeSampling.SampleCosts.values — Method

values(
    sc::InformativeSampling.SampleCosts.SampleCost,
    loc
) -> Tuple{Any, Any, Float64, Float64}

Returns the values to be used to calculate the sample cost (belief mean, standard deviation, travel distance, sample proximity).

Each concrete subtype of SampleCost needs to implement.

This can be a useful function to inspect values during optimization.

Paths.jl

InformativeSampling.Paths — Module

This module contains types and functions for searching for paths on a 2D grid using A*. Its main use is to get the path cost (distance), but it can return the full path as well.

Main public types and functions:

PathCost

finalOrientation

getPath

InformativeSampling.Paths.PathCost — Type

Struct for PathCost function data. Previous computations are kept track of in its data. Can be used multiple times for the same start cell, saving computation.

The cells of the costMatrix contain the distance to them from the start cell. NaN is a placeholder meaning no path has been calculated to that cell yet. Inf means that cell is not reachable from the start cell.

Fields:

start::Any

costMatrix::Any

resolution::Any

diffs::Any

dist::Any

heuristic::Any

frontier::Any

InformativeSampling.Paths.PathCost — Method

PathCost(
    start,
    occupancy,
    resolution;
    diagonals,
    dist,
    heuristic
) -> InformativeSampling.Paths.PathCost

The constructor initializes the path search algorithm, created for each new start cell.

Inputs:

start: the starting cell in the matrix

occupancy: a matrix indicating which cells are occupied

resolution: a vector of the width and height of each cell

diagonals: whether to include connectivity to diagonal cells or not (default true)

dist: function to compute distance from current to next cell (default weightedEuclidian)

heuristic: function to compute estimated distance from current to goal cell (default weightedEuclidian)

Note: this type and associated methods can in fact be used with any N-dimensional cost array, not just a matrix.

InformativeSampling.Paths.PathCost — Method

Search for a path to a desired goal cell using the A* algorithm.

Returns the path cost, which will be Inf if it is unreachable.

If the cells of the path are desired, use the getPath function.

InformativeSampling.Paths.finalOrientation — Method

finalOrientation(
    S::InformativeSampling.Paths.PathCost,
    goal
) -> Any

Given a PathCost and a goal point, this function returns the angle of the direction from penultimate cell to goal cell, effectively the direction at the end of the path to the goal.

InformativeSampling.Paths.getPath — Method

getPath(
    S::InformativeSampling.Paths.PathCost,
    goal
) -> Vector

Given a PathCost and a goal point, this function returns the entire list of cells from the start to the goal. Only useful to be called after the PathCost object is called with the goal cell.

ROSInterface.jl

InformativeSampling.ROSInterface — Module

This module contains the interface for passing data to and from other ROS nodes. It sets up an informative_sampling node and provides methods to handle the data. This is designed specifically for communication with Swagbot.

Main public types and functions:

ROSSampler

InformativeSampling.ROSInterface.ROSSampler — Type

A struct that stores information for communicating with Swagbot.

Objects of this type can be used as samplers in missions, meaning they can be called with a SampleInput to return its value. This object also has a length, which is the length of the number of its subscriptions and can be iterated over to get the name of each one.

Fields:

data_topics::Vector{T} where T<:Union{String, Tuple{String, String}}: vector of topic names that will be subscribed to to receive measurements

done_topic::Any: topic name that publishes a message to signify the traveling is done

pub_topic::Any: the publisher topic name

publisher::Any: the publisher topic object, created automatically from a given name

InformativeSampling.ROSInterface.ROSSampler — Method

ROSSampler(
    data_topics,
    done_topic,
    pub_topic
) -> InformativeSampling.ROSInterface.ROSSampler

Creating a ROSSampler object requires a vector of topics to subscribe to for measurement data, essentially the list of sensors onboard the robot to listen to. Each element of this list should be a 2-tuple of topics that will transmit the value and error for each sensor. This constructor initializes a ROS node and sets up a publisher to pub_topic.

InformativeSampling.ROSInterface.ROSSampler — Method

function (R::ROSSampler{String})(new_index::SampleInput)

Returns a single value from the sample location of the chosen quantity. It does this by first publishing the next location to sample. Once the location is sampled, it calls out to each topic in sequence and waits for its message.

Currently will be unused.

function (R::ROSSampler{NTuple{2, String}})(new_index::SampleInput)

Returns a single value and its error from the sample location of the chosen quantity. It does this by first publishing the next location to sample. Once the location is sampled, it calls out to each topic in sequence and waits for its message.

Currently will be unused.

InformativeSampling.ROSInterface.ROSSampler — Method

function (R::ROSSampler{String})(new_loc::Location)

Returns a vector of values from the sample location, one for each sensor measurement available. It does this by first publishing the next location to sample. Once the location is sampled, it calls out to each topic in sequence and waits for its message.

Examples

data_topics = [
    "/value1",
    "/value2"
]

done_topic = "sortie_finished"
pub_topic = "latest_sample"

sampler = ROSSampler(data_topics, done_topic, pub_topic)

location = [.1, .3]
[value1, value2] = sampler(location)

function (R::ROSSampler{NTuple{2, String}})(new_loc::Location)

Returns a vector of (value, error) pairs from the sample location, one for each sensor measurement available. It does this by first publishing the next location to sample. Once the location is sampled, it calls out to each topic in sequence and waits for its message.

Examples

data_topics = [
    ("/value1", "/error1"),
    ("/value2", "/error2")
]

done_topic = "sortie_finished"
pub_topic = "latest_sample"

sampler = ROSSampler(data_topics, done_topic, pub_topic)

location = [.1, .3]
[(value1, error1), (value2, error2)] = sampler(location)

InformativeSampling.ROSInterface.publishNextLocation — Method

publishNextLocation(publisher, new_loc::Vector{Float64})

Internal function used to send the next location to Swagbot.

LamLam, C Q (2008) Sequential adaptive designs in computer experiments for response surface model fit (Doctoral dissertation). The Ohio State University.

LiuLiu H, Cai J, Ong Y (2017) An adaptive sampling approach for kriging metamodeling by maximizing expected prediction error. Comput Chem Eng 106:171–182